+++ src/frontend.rs
@@ -40,7 +40,8 @@ pub fn compile_ron_definition(input_string : &str, options_opt : Option<CompileO
 		Ok(mut definition) =>
 		{
 			assert_eq!(definition.version, (0, 0, 1));
-			crate::rust_wgpu_backend::explicate_scheduling::explicate_scheduling(&mut definition.program);
+			definition.program = crate::rust_wgpu_backend::explicate_scheduling::
+				explicate_scheduling(definition.program);
 			ir::validation::validate_program(& definition.program);
 			let mut codegen = crate::rust_wgpu_backend::codegen::CodeGen::new(& definition.program);
 			if let Some(options) = options_opt
@@ -64,7 +65,8 @@ pub fn explicate_ron_definition(input_string : &str, options : Option<CompileOpt
 		Ok(mut definition) =>
 		{
 			assert_eq!(definition.version, (0, 0, 1));
-			crate::rust_wgpu_backend::explicate_scheduling::explicate_scheduling(&mut definition.program);
+			definition.program = crate::rust_wgpu_backend::explicate_scheduling::
+				explicate_scheduling(definition.program);
 			let output_string_result = ron::ser::to_string_pretty(& definition, pretty);
 			Ok(output_string_result.unwrap())
 		}
